+++ src/@locale/demo/metrics/index.js
@@ -1,10 +1,12 @@
 import barList     from './barList' ;
 import categoryBar from './categoryBar' ;
+import tracker     from './tracker' ;
 
 const metrics =
 {
     barList ,
     categoryBar ,
+    tracker ,
 } ;
 
 export default metrics ;